Develop

+++ www/core/Model.php
@@ -3,6 +3,7 @@
 namespace Core;
 
 use Exception;
+use InvalidArgumentException;
 use ReflectionNamedType;
 
 abstract class Model
@@ -119,7 +120,8 @@ abstract class Model
         $columns = array_keys($data);
         $queryPrepared = $pdo->prepare("INSERT INTO " . $instance->table . " (" . implode(",", $columns) . ") 
                             VALUES (:" . implode(",:", $columns) . ")");
-        return $queryPrepared->execute($data);
+        
+        return $queryPrepared->execute($data); 
     }
 
     public static function update(string $id, array $data): bool
@@ -131,8 +133,8 @@ abstract class Model
             $data = array_intersect_key($data, array_flip($instance->fillable));
         }
         $columns = [];
-        foreach ($data as $key => $value){
-            if(is_bool($value))
+        foreach ($data as $key => $value) {
+            if (is_bool($value))
                 $data[$key] = (int)$value;
             $columns[] = $key . "=:" . $key;
         }
@@ -141,12 +143,21 @@ abstract class Model
         return $queryPrepared->execute($data);
     }
 
+    public static function normalizeTableName($tableName) {
+        $tableName = explode(".", $tableName)[1];
+        $tableName = strtolower(
+            preg_replace('/([a-z])([A-Z])/', '$1_$2', $tableName)
+        );        
+        return $tableName;
+    }
     public static function findBy(string $column, string $value): ?Model
     {
 
         $dbConnector = DBConnector::getInstance();
         $pdo = $dbConnector->getPDO();
         $instance = new static();
+        if (!preg_match('/^[a-zA-Z0-9_\-]+$/', $column))
+            throw new InvalidArgumentException("Seuls les caracères [a-zA-Z0-9_\-] sont autorisé en nom de column");
 
         $queryPrepared = $pdo->prepare("SELECT * FROM " . $instance->table . " WHERE " . $column . " = :value");
         $queryPrepared->bindValue(":value", $value);
@@ -175,6 +186,25 @@ abstract class Model
         }, $results);
     }
 
+    public static function findAllBy(string $column, string $value): array
+    {
+        $dbConnector = DBConnector::getInstance();
+        $pdo = $dbConnector->getPDO();
+        if (!preg_match('/^[a-zA-Z0-9_\-]+$/', $column))
+            throw new InvalidArgumentException("Seuls les caracères [a-zA-Z0-9_\-] sont autorisé en nom de column");
+
+        $instance = new static();
+        $req = $pdo->prepare("SELECT * FROM " . $instance->table . " WHERE $column = :value");
+        $req->execute([
+            ':value' => $value
+        ]);
+        $results = $req->fetchAll();
+
+        return array_map(function ($result) {
+            return static::hydrate($result);
+        }, $results);
+    }
+
     public static function findOne(int $id): ?Model
     {
         $dbConnector = DBConnector::getInstance();
@@ -190,18 +220,60 @@ abstract class Model
         }
         return null;
     }
+    public function getTable(): string {
+        $str = $this->table;
+        $str = preg_replace('/.*\./', '', $this->table);
+        $str[0] = strtolower($str[0]);
 
+        $str = preg_replace_callback('/[A-Z]/',function($matches){
+            return '_' . strtolower($matches[0]);
+        } , $str);
+        return $str;
+    }
+
+    public function setTable(string $tableName): void {
+        $this->table = $tableName;
+    }
     public static function delete(string $attribute, mixed $value): bool
+    {
+        $dbConnector = DBConnector::getInstance();
+        $pdo = $dbConnector->getPDO();
+        if (!preg_match('/^[a-zA-Z0-9_\-]+$/', $attribute))
+            throw new InvalidArgumentException("Seuls les caracères [a-zA-Z0-9_\-] sont autorisé en nom de column");
+
+        $instance = new static();
+
+        $queryPrepared = $pdo->prepare("DELETE FROM " . $instance->table . " WHERE " . $attribute . " = " . ":" . $attribute);
+        $queryPrepared->bindValue($attribute, $value);
+        
+        $queryPrepared->execute();
+
+        return ($queryPrepared->rowCount() > 0);
+    }
+
+    public static function count()
     {
         $dbConnector = DBConnector::getInstance();
         $pdo = $dbConnector->getPDO();
 
         $instance = new static();
-        $queryPrepared = $pdo->prepare("DELETE FROM " . $instance->table . " WHERE " . $attribute . " = " . ":" . $attribute);
+        $queryPrepared = $pdo->prepare("SELECT COUNT(*) FROM " . $instance->table);
+        $queryPrepared->execute();
+        return $queryPrepared->fetchColumn();
+    }
+
+    public static function countBy(string $attribute, mixed $value,  #[SqlOperator] string $operator = "="): int
+    {
+        $dbConnector = DBConnector::getInstance();
+        $pdo = $dbConnector->getPDO();
+        if (!preg_match('/^[a-zA-Z0-9_\-]+$/', $attribute))
+            throw new InvalidArgumentException("Seuls les caracères [a-zA-Z0-9_\-] sont autorisé en nom de column");
+
+        $instance = new static();
+        $queryPrepared = $pdo->prepare("SELECT COUNT(*) FROM " . $instance->table . " WHERE " . $attribute . $operator . ":" . $attribute);
         $queryPrepared->bindValue($attribute, $value);
         $queryPrepared->execute();
-
-        return ($queryPrepared->rowCount() > 0);
+        return $queryPrepared->fetchColumn();
     }
